+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 5.1.4
+
+### Changed
+* Apply water mask to coherence coefficient first. 
+  then produce the validity mask with the masked coherence coefficient only.
+  do not use pwr when creating the validity mask. 
+
+
 ## 5.1.3
 
 ### Security